+++ ft_bzero.c
@@ -6,15 +6,15 @@
 /*   By: dximenez <dximenez@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/09 13:43:27 by dximenez          #+#    #+#             */
-/*   Updated: 2024/01/09 13:50:50 by dximenez         ###   ########.fr       */
+/*   Updated: 2024/01/09 14:35:56 by dximenez         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-void	ft_bzero(void *s, unsigned long n)
+void	ft_bzero(void *s, size_t n)
 {
-	unsigned long	i;
+	size_t	i;
 
 	i = 0;
 	while (i < n)
